feat: setup new email rendering system
