+++ app/emails/providers/mailgun-provider.ts
@@ -2,7 +2,7 @@ import formData from 'form-data';
 import Mailgun from 'mailgun.js';
 import type { IMailgunClient } from 'mailgun.js/Interfaces';
 
-import { isValidEmail } from '../validators/email.ts';
+import { isValidEmail } from '../utils/email.ts';
 import type { Email, EmailProvider } from './provider.ts';
 
 export class MailgunProvider implements EmailProvider {
