+++ app/emails/providers/mailpit-provider.ts
@@ -1,6 +1,6 @@
 import nodemailer from 'nodemailer';
 
-import { isValidEmail } from '../validators/email.ts';
+import { isValidEmail } from '../utils/email.ts';
 import type { Email, EmailProvider } from './provider.ts';
 
 export class MailpitProvider implements EmailProvider {
